# specs/004-database-setup/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Evolution API - Database Layer
  description: RESTful API for todo management with PostgreSQL database integration
  version: 1.0.0
  contact:
    name: Todo Evolution Project
    url: https://github.com/iamKhan79690/EVOLUTION-OF-TODO
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://your-app.your-domain.com
    description: Production server

tags:
  - name: health
    description: Health check and monitoring endpoints
  - name: tasks
    description: Task management operations

paths:
  /api/v1/health:
    get:
      tags:
        - health
      summary: Check API health status
      description: Returns the health status of the API and database connectivity
      operationId: health_check
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/health/database:
    get:
      tags:
        - health
      summary: Check database connectivity
      description: Verifies database connection and returns detailed status
      operationId: database_health_check
      responses:
        '200':
          description: Database connection successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseHealthResponse'
        '503':
          description: Database connection failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/health/pool:
    get:
      tags:
        - health
      summary: Get connection pool statistics
      description: Returns current connection pool usage and statistics
      operationId: pool_health_check
      responses:
        '200':
          description: Pool statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolHealthResponse'

  /api/{user_id}/tasks:
    get:
      tags:
        - tasks
      summary: List user's tasks
      description: Retrieve all tasks belonging to the specified user
      operationId: list_tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (UUID format)
          schema:
            type: string
            format: uuid
        - name: completed
          in: query
          required: false
          description: Filter by completion status
          schema:
            type: boolean
        - name: priority
          in: query
          required: false
          description: Filter by priority level
          schema:
            type: string
            enum: [low, medium, high]
        - name: tags
          in: query
          required: false
          description: Filter by tags (comma-separated)
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Number of results to skip for pagination
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                    description: Total number of tasks matching filters
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - User ID does not match JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - tasks
      summary: Create new task
      description: Create a new task for the specified user
      operationId: create_task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (UUID format)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - JWT token missing or invalid
        '403':
          description: Forbidden - User ID does not match JWT
        '404':
          description: User not found

  /api/{user_id}/tasks/{task_id}:
    get:
      tags:
        - tasks
      summary: Get specific task
      description: Retrieve details of a specific task
      operationId: get_task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (UUID format)
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          description: Task ID (UUID format)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Task does not belong to user
        '404':
          description: Task not found

    put:
      tags:
        - tasks
      summary: Update task
      description: Update an existing task
      operationId: update_task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Task does not belong to user
        '404':
          description: Task not found

    delete:
      tags:
        - tasks
      summary: Delete task
      description: Delete a specific task
      operationId: delete_task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Task does not belong to user
        '404':
          description: Task not found

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      tags:
        - tasks
      summary: Toggle task completion
      description: Mark task as complete or incomplete
      operationId: toggle_task_completion
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: New completion status (optional - toggles if not provided)
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Task does not belong to user
        '404':
          description: Task not found

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User unique identifier
        email:
          type: string
          format: email
          description: User email address
        name:
          type: string
          description: User display name
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Task unique identifier
        user_id:
          type: string
          format: uuid
          description: Owner user ID
        title:
          type: string
          maxLength: 500
          description: Task title
        description:
          type: string
          description: Task description (optional)
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          description: Task priority level
        tags:
          type: array
          items:
            type: string
          description: Task tags
        due_date:
          type: string
          format: date-time
          description: Task due date (optional)
        is_recurring:
          type: boolean
          default: false
          description: Whether task is recurring
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly, yearly]
          description: Recurrence pattern (if recurring)
        recurrence_end_date:
          type: string
          format: date-time
          description: End date for recurrence (optional)
        completed:
          type: boolean
          default: false
          description: Task completion status
        completed_at:
          type: string
          format: date-time
          description: Completion timestamp (set when completed)
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        user:
          $ref: '#/components/schemas/User'
      required:
        - id
        - user_id
        - title
        - priority
        - completed
        - created_at
        - updated_at

    TaskCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
          description: Task title
          example: "Complete project documentation"
        description:
          type: string
          description: Task description (optional)
          example: "Write comprehensive documentation for the new feature"
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          description: Task priority level
          example: "high"
        tags:
          type: array
          items:
            type: string
          description: Task tags
          example: ["documentation", "urgent"]
        due_date:
          type: string
          format: date-time
          description: Task due date (optional)
          example: "2025-12-15T10:00:00Z"
        is_recurring:
          type: boolean
          default: false
          description: Whether task is recurring
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly, yearly]
          description: Recurrence pattern (if recurring)
          example: "weekly"
        recurrence_end_date:
          type: string
          format: date-time
          description: End date for recurrence (optional)
      required:
        - title

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
          description: Updated task title
        description:
          type: string
          description: Updated task description
        priority:
          type: string
          enum: [low, medium, high]
          description: Updated priority level
        tags:
          type: array
          items:
            type: string
          description: Updated task tags
        due_date:
          type: string
          format: date-time
          description: Updated due date
        is_recurring:
          type: boolean
          description: Updated recurring status
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly, yearly]
          description: Updated recurrence pattern
        recurrence_end_date:
          type: string
          format: date-time
          description: Updated recurrence end date

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Overall API health status
        timestamp:
          type: string
          format: date-time
          description: Health check timestamp
        version:
          type: string
          description: API version

    DatabaseHealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Database connection status
        database:
          type: string
          enum: [connected, disconnected]
          description: Database connectivity status
        timestamp:
          type: string
          format: date-time
          description: Health check timestamp
        response_time_ms:
          type: number
          description: Database query response time in milliseconds

    PoolHealthResponse:
      type: object
      properties:
        pool_size:
          type: integer
          description: Current pool size
        checked_in:
          type: integer
          description: Number of connections checked in
        checked_out:
          type: integer
          description: Number of connections checked out
        overflow:
          type: integer
          description: Number of overflow connections
        invalid:
          type: integer
          description: Number of invalid connections
        timestamp:
          type: string
          format: date-time
          description: Statistics timestamp

    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
      required:
        - detail

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Better Auth

security:
  - BearerAuth: []

# Response examples
examples:
  TaskCreated:
    value:
      id: "550e8400-e29b-41d4-a716-446655440000"
      user_id: "550e8400-e29b-41d4-a716-446655440001"
      title: "Complete project documentation"
      description: "Write comprehensive documentation for the new feature"
      priority: "high"
      tags: ["documentation", "urgent"]
      due_date: "2025-12-15T10:00:00Z"
      is_recurring: false
      completed: false
      created_at: "2025-12-07T10:30:00Z"
      updated_at: "2025-12-07T10:30:00Z"

  HealthCheck:
    value:
      status: "healthy"
      database: "connected"
      timestamp: "2025-12-07T10:30:00Z"
      response_time_ms: 45

  ErrorUnauthorized:
    value:
      detail: "Could not validate credentials"
      status_code: 401
      timestamp: "2025-12-07T10:30:00Z"